requirements.txt
micropython-rp2-stubs
adafruit-ampy
mpy-cross
watchdog
setuptools
